requirements.txt
flask
requests
gunicorn
gevent
qiniu
wechatpy
pycryptodome
DBUtils
lxml
pymysql
tencentcloud-sdk-python